% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{bkfee}
\alias{bkfee}
\title{Compute fee in a block}
\usage{
bkfee(con, height)
}
\arguments{
\item{con}{\code{CONRPC}, configuration object.}

\item{height}{\code{integer}, the height of the block.}
}
\value{
\code{numeric}
}
\description{
This function returns the fee of the coinbase transaction.
Hereby, the mining reward has been deducted.
Initially, the mining reward was 50 BTC and is halved every
210,000 blocks.
}
\seealso{
Other UtilityFuncs: \code{\link{blockattime}},
  \code{\link{blockstats}}, \code{\link{date2int}},
  \code{\link{int2date}}, \code{\link{intMaxDay}},
  \code{\link{intMinDay}}, \code{\link{intRangeDay}},
  \code{\link{intRangePeriod}}, \code{\link{timeofblock}},
  \code{\link{txfee}}, \code{\link{txids}},
  \code{\link{txinids}}, \code{\link{txstats}},
  \code{\link{utxoage}}, \code{\link{utxotype}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
