% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{blockattime}
\alias{blockattime}
\title{Block height at time}
\usage{
blockattime(con, targetdate)
}
\arguments{
\item{con}{\code{CONRPC}, configuration object.}

\item{targetdate}{\code{POSIXct}, the date/time of closest block heights.}
}
\value{
\code{data.frame}: the heights, the times and
the time differences (in minutes) to the provided date/time.
}
\description{
This function returns the block heights closest to
a provided date/time (time zone is GMT).
}
\seealso{
Other UtilityFuncs: \code{\link{blockstats}},
  \code{\link{date2int}}, \code{\link{int2date}},
  \code{\link{intMaxDay}}, \code{\link{intMinDay}},
  \code{\link{intRangeDay}}, \code{\link{intRangePeriod}},
  \code{\link{timeofblock}}, \code{\link{txfee}},
  \code{\link{txids}}, \code{\link{txinids}},
  \code{\link{txstats}}, \code{\link{utxoage}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
