% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{blockstats}
\alias{blockstats}
\title{Obtaining statistics of a block}
\usage{
blockstats(con, height, excoinbase = TRUE)
}
\arguments{
\item{con}{\code{CONRPC}, configuration object.}

\item{height}{\code{integer}, the block's height.}

\item{excoinbase}{\code{logical}, whether coinbase transaction
should be excluded (default is \code{TRUE}).}
}
\value{
An object of class \code{data.frame}
}
\description{
This function returns key statistics of a block's content,
such as the time, the count of transactions,
and summary statistics of the UTXOs.
}
\seealso{
Other UtilityFuncs: \code{\link{bkfee}},
  \code{\link{blockattime}}, \code{\link{date2int}},
  \code{\link{int2date}}, \code{\link{intMaxDay}},
  \code{\link{intMinDay}}, \code{\link{intRangeDay}},
  \code{\link{intRangePeriod}}, \code{\link{timeofblock}},
  \code{\link{txfee}}, \code{\link{txids}},
  \code{\link{txinids}}, \code{\link{txstats}},
  \code{\link{utxoage}}, \code{\link{utxotype}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
