% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CharacterOrList-class}
\alias{CharacterOrList-class}
\title{S4 Class Union character or list}
\description{
S4-class union of \code{character} or \code{list}.
}
\seealso{
Other bitcoind functions: \code{\link{ANSRPC-class}},
  \code{\link{CONRPC-class}},
  \code{\link{NullOrCharacter-class}},
  \code{\link{NullOrInteger-class}}, \code{\link{conrpc}},
  \code{\link{rpcpost}}, \code{\link{startbtc}},
  \code{\link{stopbtc}}
}
