% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CONRPC}
\alias{CONRPC}
\alias{CONRPC-class}
\title{The CONRPC class}
\description{
S4-class for curl connections to RPC-JSON.
}
\details{
The slots \code{rpcuser} and \code{rpcpassowrd} are required in the call
to \code{curl}. Furthermore, the fully qualified path to bitcoin.conf is
required for starting and stopping bitcoind as daemon.
}
\author{
Bernhard Pfaff
}
