% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BtcAddresses.R
\name{createBtcAdr}
\alias{createBtcAdr}
\title{Create BTC addresses}
\usage{
createBtcAdr(privkey, mainnet = TRUE)
}
\arguments{
\item{privkey}{\code{character}, a private key.}

\item{mainnet}{\code{logical}, for which net the keys should belong to.}
}
\value{
Object of S4-class \code{BTCADR}
}
\description{
This function creates an object of S4-class \code{BTCADR}.
}
\references{
\url{https://en.bitcoin.it/wiki/Address}
}
\seealso{
Other BtcAdresses: \code{\link{BTCADR-class}},
  \code{\link{PrivKey2PubKey}}, \code{\link{PrivKey2Wif}},
  \code{\link{PubHash2BtcAdr}},
  \code{\link{PubKey2PubHash}}, \code{\link{Wif2PrivKey}},
  \code{\link{base58CheckDecode}},
  \code{\link{base58CheckEncode}}, \code{\link{concatHex}},
  \code{\link{createPrivateKey}}, \code{\link{decodeHex}},
  \code{\link{hash160}}, \code{\link{hash256}},
  \code{\link{validBtcAdr}}
}
\author{
Bernhard Pfaff
}
\concept{BtcAdresses}
