% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawTransactions.R
\name{decoderawtransaction}
\alias{decoderawtransaction}
\title{RPC-JSON API: decoderawtransaction}
\usage{
decoderawtransaction(con, hexstring)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}

\item{hexstring}{\code{character}, the transaction hex string.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Return a JSON object representing the serialized,
hex-encoded transaction.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getblock},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other RawTransactions RPCs: \code{\link{getrawtransaction}}
}
\author{
Bernhard Pfaff
}
\concept{RawTransactions RPCs}
