% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ECPARAM-class}
\alias{ECPARAM-class}
\title{The ECPARAM class}
\description{
S4-class for elliptic curve parameters. 
Objects of this class do contain the big integer parameters of
elliptic curves. Instances of this class are ordinarily created
by a call to \code{exparam}
}
\section{Slots}{

\describe{
\item{\code{p}}{\code{bigz}, curve dimension.}

\item{\code{a}}{\code{bigz}, parameter.}

\item{\code{b}}{\code{bigz}, parameter.}
}}

\references{
\url{https://en.bitcoin.it/wiki/Secp256k1}
}
\seealso{
Other EllipticCurve: \code{\link{ECPOINT-class}},
  \code{\link{EcparamOrNull-class}},
  \code{\link{containsPoint}}, \code{\link{ecoperators}},
  \code{\link{ecparam}}, \code{\link{ecpoint}},
  \code{\link{isNull}}
}
\author{
Bernhard Pfaff
}
\concept{EllipticCurve}
