% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EcparamOrNull-class}
\alias{EcparamOrNull-class}
\title{S4 Class Union ECPARAM or NULL}
\description{
S4-class union of \code{NULL} or \code{ECPARAM}.
}
\references{
\url{https://en.bitcoin.it/wiki/Secp256k1}
}
\seealso{
Other EllipticCurve: \code{\link{ECPARAM-class}},
  \code{\link{ECPOINT-class}}, \code{\link{containsPoint}},
  \code{\link{ecoperators}}, \code{\link{ecparam}},
  \code{\link{ecpoint}}, \code{\link{isNull}}
}
\author{
Bernhard Pfaff
}
\concept{EllipticCurve}
