% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ECPOINT-class}
\alias{ECPOINT-class}
\title{S4 Class ECPOINT}
\description{
S4-class for a point on an elliptic curve.
Ordinarily, objects are created by calling
\code{ecpoint}.
}
\section{Slots}{

\describe{
\item{\code{ecparam}}{\code{ECPARAM}}

\item{\code{x}}{\code{bigz}}

\item{\code{y}}{\code{bigz}}

\item{\code{r}}{\code{bigz}}
}}

\references{
\url{https://en.bitcoin.it/wiki/Secp256k1}
}
\seealso{
Other EllipticCurve: \code{\link{ECPARAM-class}},
  \code{\link{EcparamOrNull-class}},
  \code{\link{containsPoint}}, \code{\link{ecoperators}},
  \code{\link{ecparam}}, \code{\link{ecpoint}},
  \code{\link{isNull}}
}
\author{
Bernhard Pfaff
}
\concept{EllipticCurve}
