% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkRpcs.R
\name{getaddednodeinfo}
\alias{getaddednodeinfo}
\title{RPC-JSON API: getaddednodeinfo}
\usage{
getaddednodeinfo(con, node = NULL)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}

\item{node}{\code{character} the node (see \code{getpeerinfo()}
for nodes).}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns information about the given added node,
or all added nodes (note that onetry addnodes
are not listed here)
}
\references{
\url{https://bitcoin.org/en/developer-reference#getaddednodeinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Network RPCs: \code{\link{addnode}},
  \code{\link{clearbanned}}, \code{\link{disconnectnode}},
  \code{\link{getconnectioncount}},
  \code{\link{getnettotals}}, \code{\link{getnetworkinfo}},
  \code{\link{getpeerinfo}}, \code{\link{listbanned}},
  \code{\link{ping}}, \code{\link{setnetworkactive}}
}
\author{
Bernhard Pfaff
}
\concept{Network RPCs}
