% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getbestblockhash}
\alias{getbestblockhash}
\title{RPC-JSON API: getbestblockhash}
\usage{
getbestblockhash(conobj)
}
\arguments{
\item{conobj}{object of class \code{CONRPC}.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returns the hash of the best (tip) block in the longest blockchain.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getbestblockhash},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getblockchaininfo}}
}
\author{
Bernhard Pfaff
}
