% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getblockcount}
\alias{getblockcount}
\title{RPC-JSON API: getblockcount}
\usage{
getblockcount(obj)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returns the number of blocks in the longest blockchain.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getblockcount},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockhash}}, \code{\link{getblock}}
}
\author{
Bernhard Pfaff
}
