% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getblockhash}
\alias{getblockhash}
\title{RPC-JSON API: getblockhash}
\usage{
getblockhash(obj, height)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}

\item{height}{\code{integer} the height index.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returns hash of block in best-block-chain at height provided.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getblockhash},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockcount}}, \code{\link{getblock}}
}
\author{
Bernhard Pfaff
}
