% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getchaintxstats}
\alias{getchaintxstats}
\title{RPC-JSON API: getchaintxstats}
\usage{
getchaintxstats(con, nblocks = NULL, blockhash = NULL)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}

\item{nblocks}{\code{integer} optional, size of the window in
number of blocks (default: one month).}

\item{blockhash}{\code{character} optional, the hash of the block
that ends the window.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Compute statistics about the total number and
rate of transactions in the chain.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getchaintxstats},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{decodescript}},
  \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockcount}}, \code{\link{getblockhash}},
  \code{\link{getblockheader}}, \code{\link{getblock}},
  \code{\link{getchaintips}}, \code{\link{getdifficulty}},
  \code{\link{getmempoolancestors}},
  \code{\link{getmempooldescendants}},
  \code{\link{getmempoolentry}},
  \code{\link{getmempoolinfo}},
  \code{\link{getrawmempool}}, \code{\link{gettxoutproof}},
  \code{\link{gettxoutsetinfo}}, \code{\link{gettxout}},
  \code{\link{preciousblock}},
  \code{\link{pruneblockchain}}, \code{\link{verifychain}},
  \code{\link{verifytxoutproof}}
}
\author{
Bernhard Pfaff
}
\concept{Blockchain RPCs}
