% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkRpcs.R
\name{getconnectioncount}
\alias{getconnectioncount}
\title{RPC-JSON API: getconnectioncount}
\usage{
getconnectioncount(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns the number of connections to other nodes.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getconnectioncount},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Network RPCs: \code{\link{addnode}},
  \code{\link{clearbanned}}, \code{\link{disconnectnode}},
  \code{\link{getaddednodeinfo}},
  \code{\link{getnettotals}}, \code{\link{getnetworkinfo}},
  \code{\link{getpeerinfo}}, \code{\link{listbanned}},
  \code{\link{ping}}, \code{\link{setnetworkactive}}
}
\author{
Bernhard Pfaff
}
\concept{Network RPCs}
