% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethelp.R
\name{gethelp}
\alias{gethelp}
\title{RPC-JSON API: Help}
\usage{
gethelp(conobj, rpcname = "")
}
\arguments{
\item{conobj}{object of class \code{CONRPC}.}

\item{rpcname}{\code{character}, name of RPC function.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returning information about RPC functions.
}
\seealso{
Other Control RPCs: \code{\link{getinfo}}
}
\author{
Bernhard Pfaff
}
