% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRpcs.R, R/getinfo.R
\name{getinfo}
\alias{getinfo}
\alias{getinfo}
\title{RPC-JSON API: getinfo}
\usage{
getinfo(conobj)

getinfo(conobj)
}
\arguments{
\item{conobj}{object of class \code{CONRPC}.}

\item{conobj}{object of class \code{CONRPC}.}
}
\value{
A coerced \code{list} object from RPC-JSON API.

A coerced \code{list} object from RPC-JSON API.
}
\description{
Returning information about bitcoin configuration and settings.

Returning information about bitcoin configuration and settings.
}
\details{
WARNING: getinfo is deprecated and will be fully removed in 0.16.
Projects should transition to using getblockchaininfo, getnetworkinfo, and
getwalletinfo before upgrading to 0.16.

WARNING: getinfo is deprecated and will be fully removed in 0.16.
Projects should transition to using getblockchaininfo, getnetworkinfo, and
getwalletinfo before upgrading to 0.16.
}
\references{
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}

\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Control RPCs: \code{\link{gethelp}}

Other Control RPCs: \code{\link{gethelp}}
}
\author{
Bernhard Pfaff

Bernhard Pfaff
}
