% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getmempoolentry}
\alias{getmempoolentry}
\title{RPC-JSON API: getmempoolentry}
\usage{
getmempoolentry(obj, txid)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}

\item{txid}{\code{character}, the transaction id (must be in mempool).}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returns mempool data for given transaction.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getmempoolentry},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockcount}}, \code{\link{getblockhash}},
  \code{\link{getblockheader}}, \code{\link{getblock}},
  \code{\link{getchaintips}},
  \code{\link{getchaintxstats}},
  \code{\link{getdifficulty}},
  \code{\link{getmempoolinfo}},
  \code{\link{gettxoutsetinfo}},
  \code{\link{preciousblock}},
  \code{\link{pruneblockchain}}
}
\author{
Bernhard Pfaff
}
