% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkRpcs.R
\name{getnettotals}
\alias{getnettotals}
\title{RPC-JSON API: getnettotals}
\usage{
getnettotals(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns information about network traffic, including bytes in,
bytes out, and current time.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getnettotals},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Network RPCs: \code{\link{addnode}},
  \code{\link{clearbanned}}, \code{\link{disconnectnode}},
  \code{\link{getaddednodeinfo}},
  \code{\link{getconnectioncount}},
  \code{\link{getnetworkinfo}}, \code{\link{getpeerinfo}},
  \code{\link{listbanned}}, \code{\link{ping}},
  \code{\link{setnetworkactive}}
}
\author{
Bernhard Pfaff
}
\concept{Network RPCs}
