% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRpcs.R
\name{getnetworkinfo}
\alias{getnetworkinfo}
\title{RPC-JSON API: getnetworkinfo}
\usage{
getnetworkinfo(obj)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns an object containing various state info regarding P2P networking.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getnetworkinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Control RPCs: \code{\link{gethelp}},
  \code{\link{getinfo}}, \code{\link{getwalletinfo}}
}
\author{
Bernhard Pfaff
}
