% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkRpcs.R
\name{getpeerinfo}
\alias{getpeerinfo}
\title{RPC-JSON API: getpeerinfo}
\usage{
getpeerinfo(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns data about each connected network node as a json array of objects.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getpeerinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Network RPCs: \code{\link{addnode}},
  \code{\link{clearbanned}}, \code{\link{disconnectnode}},
  \code{\link{getaddednodeinfo}},
  \code{\link{getconnectioncount}},
  \code{\link{getnettotals}}, \code{\link{getnetworkinfo}},
  \code{\link{listbanned}}, \code{\link{ping}},
  \code{\link{setnetworkactive}}
}
\author{
Bernhard Pfaff
}
\concept{Network RPCs}
