% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getrawmempool}
\alias{getrawmempool}
\title{RPC-JSON API: getrawmempool}
\usage{
getrawmempool(obj, verbose = TRUE)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}

\item{verbose}{\code{logical}, \code{TRUE} for a json object,
\code{FALSE} for array of transaction ids}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns all transaction ids in memory pool as a json array
of string transaction ids.
Hint: use getmempoolentry to fetch a specific transaction
from the mempool.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getrawmempool},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockcount}}, \code{\link{getblockhash}},
  \code{\link{getblockheader}}, \code{\link{getblock}},
  \code{\link{getchaintips}},
  \code{\link{getchaintxstats}},
  \code{\link{getdifficulty}},
  \code{\link{getmempoolancestors}},
  \code{\link{getmempooldescendants}},
  \code{\link{getmempoolentry}},
  \code{\link{getmempoolinfo}},
  \code{\link{gettxoutproof}},
  \code{\link{gettxoutsetinfo}}, \code{\link{gettxout}},
  \code{\link{preciousblock}},
  \code{\link{pruneblockchain}}, \code{\link{verifychain}},
  \code{\link{verifytxoutproof}}
}
\author{
Bernhard Pfaff
}
