% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawTransactions.R
\name{getrawtransaction}
\alias{getrawtransaction}
\title{RPC-JSON API: getrawtransaction}
\usage{
getrawtransaction(con, txid, verbose = FALSE)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}

\item{txid}{\code{character}, the transaction id.}

\item{verbose}{\code{logical}, type of output.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns the raw transaction data.
}
\section{}{
 Details
By default this function only works for mempool transactions.
If the -txindex option is enabled, it also works for blockchain transactions.
DEPRECATED: for now, it also works for transactions with unspent outputs.
If verbose is 'true', returns an object with information about 'txid'.
If verbose is 'false' or omitted, returns a string that is serialized,
hex-encoded data for 'txid'.
}

\references{
\url{https://bitcoin.org/en/developer-reference#getblock},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other RawTransactions RPCs: \code{\link{decoderawtransaction}}
}
\author{
Bernhard Pfaff
}
\concept{RawTransactions RPCs}
