% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{gettxoutsetinfo}
\alias{gettxoutsetinfo}
\title{RPC-JSON API: gettxoutsetinfo}
\usage{
gettxoutsetinfo(obj)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returns statistics about the unspent transaction output set.
Note this call may take some time.
}
\references{
\url{https://bitcoin.org/en/developer-reference#gettxoutsetinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockcount}}, \code{\link{getblockhash}},
  \code{\link{getblockheader}}, \code{\link{getblock}},
  \code{\link{getchaintips}},
  \code{\link{getchaintxstats}},
  \code{\link{getdifficulty}},
  \code{\link{getmempoolentry}},
  \code{\link{getmempoolinfo}},
  \code{\link{preciousblock}},
  \code{\link{pruneblockchain}}
}
\author{
Bernhard Pfaff
}
