% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRpcs.R
\name{getwalletinfo}
\alias{getwalletinfo}
\title{RPC-JSON API: getwalletinfo}
\usage{
getwalletinfo(obj)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returning information about bitcoin wallet.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getwalletinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Control RPCs: \code{\link{gethelp}},
  \code{\link{getinfo}}, \code{\link{getnetworkinfo}}
}
\author{
Bernhard Pfaff
}
