% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BtcAddresses.R
\name{hash160}
\alias{hash160}
\title{BTC hash160}
\usage{
hash160(d)
}
\arguments{
\item{d}{\code{raw}, vector.}
}
\value{
\code{character}, the value of \code{d} hashed with
\code{sha256} and \code{ripemd160}.
}
\description{
This function returns the hash by applying the \code{sha256} hashing
first and then to the resulting hash the \code{ripemd160} algorithm.
}
\references{
\url{https://en.bitcoin.it/wiki/Address}
}
\seealso{
Other BtcAdresses: \code{\link{BTCADR-class}},
  \code{\link{PrivKey2PubKey}}, \code{\link{PrivKey2Wif}},
  \code{\link{PubHash2BtcAdr}},
  \code{\link{PubKey2PubHash}}, \code{\link{Wif2PrivKey}},
  \code{\link{base58CheckDecode}},
  \code{\link{base58CheckEncode}}, \code{\link{concatHex}},
  \code{\link{createBtcAdr}},
  \code{\link{createPrivateKey}}, \code{\link{decodeHex}},
  \code{\link{hash256}}, \code{\link{validBtcAdr}}
}
\author{
Bernhard Pfaff
}
\concept{BtcAdresses}
