% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BtcAddresses.R
\name{hash256}
\alias{hash256}
\title{BTC hash256}
\usage{
hash256(d)
}
\arguments{
\item{d}{\code{raw}, vector.}
}
\value{
\code{character}, the value of \code{d} hashed twice.
}
\description{
This function returns the hash by applying the \code{sha256} hashing
algorithm twice to a \code{raw} object.
}
\references{
\url{https://en.bitcoin.it/wiki/Address}
}
\seealso{
Other BtcAdresses: \code{\link{BTCADR-class}},
  \code{\link{PrivKey2PubKey}}, \code{\link{PrivKey2Wif}},
  \code{\link{PubHash2BtcAdr}},
  \code{\link{PubKey2PubHash}}, \code{\link{Wif2PrivKey}},
  \code{\link{base58CheckDecode}},
  \code{\link{base58CheckEncode}}, \code{\link{concatHex}},
  \code{\link{createBtcAdr}},
  \code{\link{createPrivateKey}}, \code{\link{decodeHex}},
  \code{\link{hash160}}, \code{\link{validBtcAdr}}
}
\author{
Bernhard Pfaff
}
\concept{BtcAdresses}
