% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{intRangeDay}
\alias{intRangeDay}
\title{Integer range within a day}
\usage{
intRangeDay(x)
}
\arguments{
\item{x}{\code{POSIXct}, date/time object.}
}
\value{
\code{integer}
}
\description{
This function returns the associated \code{integer} times
for the start and end of a specific day.
}
\examples{
d1 <- "2017-03-15"
intRangeDay(d1)
intMinDay(d1)
intMaxDay(d1)
}
\seealso{
Other UtilityFuncs: \code{\link{date2int}},
  \code{\link{int2date}}, \code{\link{intMaxDay}},
  \code{\link{intMinDay}}, \code{\link{intRangePeriod}}
}
\author{
Bernhard Pfaff
}
