% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{intRangePeriod}
\alias{intRangePeriod}
\title{Integer range between two dates}
\usage{
intRangePeriod(d1, d2)
}
\arguments{
\item{d1}{\code{POSIXct}, date/time object.}

\item{d2}{\code{POSIXct}, date/time object.}
}
\value{
\code{integer}
}
\description{
This function returns the associated \code{integer} times
for the start of date \code{d1} and the end of date \code{d2}.
}
\examples{
d1 <- "2017-03-15"
d2 <- "2017-04-15"
intRangePeriod(d1, d2)
intMinDay(d1)
intMaxDay(d2)
}
\seealso{
Other UtilityFuncs: \code{\link{bkfee}},
  \code{\link{blockattime}}, \code{\link{blockstats}},
  \code{\link{date2int}}, \code{\link{int2date}},
  \code{\link{intMaxDay}}, \code{\link{intMinDay}},
  \code{\link{intRangeDay}}, \code{\link{timeofblock}},
  \code{\link{txfee}}, \code{\link{txids}},
  \code{\link{txinids}}, \code{\link{txstats}},
  \code{\link{utxoage}}, \code{\link{utxotype}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
