% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllipticCurves.R
\docType{methods}
\name{isNull}
\alias{isNull}
\alias{isNull,ECPOINT-method}
\title{Test for empty EC point}
\usage{
isNull(x)

\S4method{isNull}{ECPOINT}(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{logical}
}
\description{
Checks whether an EC point does exist.
}
\references{
\url{https://en.bitcoin.it/wiki/Secp256k1}
}
\seealso{
Other EllipticCurve: \code{\link{ECPARAM-class}},
  \code{\link{ECPOINT-class}},
  \code{\link{EcparamOrNull-class}},
  \code{\link{containsPoint}}, \code{\link{ecoperators}},
  \code{\link{ecparam}}, \code{\link{ecpoint}}
}
\author{
Bernhard Pfaff
}
\concept{EllipticCurve}
