% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkRpcs.R
\name{listbanned}
\alias{listbanned}
\title{RPC-JSON API: listbanned}
\usage{
listbanned(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
List all banned IPs/Subnets.
}
\references{
\url{https://bitcoin.org/en/developer-reference#listbanned},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Network RPCs: \code{\link{addnode}},
  \code{\link{clearbanned}}, \code{\link{disconnectnode}},
  \code{\link{getaddednodeinfo}},
  \code{\link{getconnectioncount}},
  \code{\link{getnettotals}}, \code{\link{getnetworkinfo}},
  \code{\link{getpeerinfo}}, \code{\link{ping}},
  \code{\link{setnetworkactive}}
}
\author{
Bernhard Pfaff
}
\concept{Network RPCs}
