% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NullOrCharacter-class}
\alias{NullOrCharacter-class}
\title{S4 Class Union NULL or character}
\description{
S4-class union of \code{NULL} or \code{character}.
}
\seealso{
Other bitcoind functions: \code{\link{ANSRPC-class}},
  \code{\link{CONRPC-class}},
  \code{\link{NullOrInteger-class}}, \code{\link{conrpc}},
  \code{\link{rpcpost}}, \code{\link{startbtc}},
  \code{\link{stopbtc}}
}
\concept{bitcoind functions}
