% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NullOrInteger-class}
\alias{NullOrInteger-class}
\title{S4 Class Union NULL or integer}
\description{
S4-class union of \code{NULL} or \code{integer}.
}
\seealso{
Other bitcoind functions: \code{\link{ANSRPC-class}},
  \code{\link{CONRPC-class}},
  \code{\link{NullOrCharacter-class}},
  \code{\link{conrpc}}, \code{\link{rpcpost}},
  \code{\link{startbtc}}, \code{\link{stopbtc}}
}
\concept{bitcoind functions}
