% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoind.R
\name{rpcpost}
\alias{rpcpost}
\title{HTTP post of RPC-JSON}
\usage{
rpcpost(conobj, api, plist = list())
}
\arguments{
\item{api}{\code{character} the name of the RPC function.}

\item{plist}{\code{list} a named list object of the parameters for \code{api}}

\item{confbtc}{\code{CONRPC} object, returned from \code{conrpc()}.}
}
\value{
A \code{list} object, coerced JSON answer from RPC.
}
\description{
This function executes an RPC-JSON post.
}
\seealso{
Other bitcoind functions: \code{\link{conrpc}},
  \code{\link{startbtc}}, \code{\link{stopbtc}}
}
\author{
Bernhard Pfaff
}
