% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,ANSRPC-method}
\title{show-methods}
\usage{
\S4method{show}{ANSRPC}(object)
}
\arguments{
\item{object}{a S4-class object.}
}
\description{
Defined \code{show}-methods for S4-classes.
}
