% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoind.R
\name{startbtc}
\alias{startbtc}
\title{Start bitcoind server process}
\usage{
startbtc(rpcuser = "rbtc", rpcpassword = "$GiveItATry4Me",
  conf.file = NULL)
}
\arguments{
\item{rpcuser}{\code{character}, the name of the RPC-user.}

\item{rpcpassword}{\code{character}, a given (strong) password.}

\item{conf.file}{\code{character}, the path and filename to the
bitcoind configuration file.}
}
\value{
\code{NULL}
}
\description{
This function does start the bitcoind-server process.
It should only be called when no suitable RPC-JSON
process is running
}
\details{
The default value for \code{conf.file} is \code{NULL}.
In this case, the bitcoind process is started with the options:
server, rpcuser and rpcpassword. The process is started by calling
\code{system2()}.
}
\author{
Bernhard Pfaff
}
