% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{timeofblock}
\alias{timeofblock}
\title{Time of a block}
\usage{
timeofblock(con, height)
}
\arguments{
\item{con}{\code{CONRPC}, configuration object.}

\item{height}{\code{integer}, the height of the block.}
}
\value{
\code{POSIXct}
}
\description{
This function returns the time of a block in GMT.
}
\seealso{
Other UtilityFuncs: \code{\link{bkfee}},
  \code{\link{blockattime}}, \code{\link{blockstats}},
  \code{\link{date2int}}, \code{\link{int2date}},
  \code{\link{intMaxDay}}, \code{\link{intMinDay}},
  \code{\link{intRangeDay}}, \code{\link{intRangePeriod}},
  \code{\link{txfee}}, \code{\link{txids}},
  \code{\link{txinids}}, \code{\link{txstats}},
  \code{\link{utxoage}}, \code{\link{utxotype}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
