% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{txfee}
\alias{txfee}
\title{Compute fee of a transaction}
\usage{
txfee(con, txid)
}
\arguments{
\item{con}{\code{CONRPC}, configuration object.}

\item{txid}{\code{character}, the id of the transaction.}
}
\value{
\code{numeric}
}
\description{
This function returns the implicit fee of a transaction,
by computing the difference between the sum of its inputs
and the sum of its outputs.
}
\seealso{
Other UtilityFuncs: \code{\link{blockattime}},
  \code{\link{blockstats}}, \code{\link{date2int}},
  \code{\link{int2date}}, \code{\link{intMaxDay}},
  \code{\link{intMinDay}}, \code{\link{intRangeDay}},
  \code{\link{intRangePeriod}}, \code{\link{timeofblock}},
  \code{\link{txids}}, \code{\link{txinids}},
  \code{\link{txstats}}, \code{\link{utxoage}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
