% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{txinids}
\alias{txinids}
\title{Retrieving the input transaction IDs}
\usage{
txinids(con, txid)
}
\arguments{
\item{con}{\code{CONRPC}, configuration object.}

\item{txid}{\code{character}, the id of the transaction.}
}
\value{
\code{data.frame}, the transaction ID(s) and
the position(s) of the previous UTXO(s).
}
\description{
This function returns the transaction IDs of the inputs for
a given transaction.
}
\seealso{
Other UtilityFuncs: \code{\link{bkfee}},
  \code{\link{blockattime}}, \code{\link{blockstats}},
  \code{\link{date2int}}, \code{\link{int2date}},
  \code{\link{intMaxDay}}, \code{\link{intMinDay}},
  \code{\link{intRangeDay}}, \code{\link{intRangePeriod}},
  \code{\link{timeofblock}}, \code{\link{txfee}},
  \code{\link{txids}}, \code{\link{txstats}},
  \code{\link{utxoage}}, \code{\link{utxotype}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
