% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{validBtcAdr}
\alias{validBtcAdr}
\title{Validate S4-class BTCADR}
\usage{
validBtcAdr(object)
}
\arguments{
\item{object}{\code{BTCADR} object}
}
\description{
This function validates objects of S4-class
\code{BTCADR}. Hereby, checks are conducted
with respect to the first character of the addresses;
their consistency with the net version and
the correspondence of the checksums.
}
\references{
\url{https://en.bitcoin.it/wiki/Address}
}
\seealso{
Other BtcAdresses: \code{\link{BTCADR-class}},
  \code{\link{PrivKey2PubKey}}, \code{\link{PrivKey2Wif}},
  \code{\link{PubHash2BtcAdr}},
  \code{\link{PubKey2PubHash}}, \code{\link{Wif2PrivKey}},
  \code{\link{base58CheckDecode}},
  \code{\link{base58CheckEncode}}, \code{\link{concatHex}},
  \code{\link{createBtcAdr}},
  \code{\link{createPrivateKey}}, \code{\link{decodeHex}},
  \code{\link{hash160}}, \code{\link{hash256}}
}
\author{
Bernhard Pfaff
}
\concept{BtcAdresses}
