% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ANSRPC-class}
\alias{ANSRPC-class}
\title{The ANSRPC class}
\description{
This class definition is employed to cast the JSON-objects
returned by API-calls to bitcoind.
}
\section{Slots}{

\describe{
\item{\code{rpcname}}{\code{character} the name of the API.}

\item{\code{result}}{\code{ANY} the output/result of the API.}

\item{\code{ecode}}{\code{NullOrInteger} the error code,
in case of no error \code{NULL}.}

\item{\code{emessage}}{\code{NullOrIntegerCharacter} the error message,
in case of no error \code{NULL}.}

\item{\code{id}}{\code{character} identifier to API-call.}
}}

\seealso{
Other bitcoind functions: \code{\link{CONRPC-class}},
  \code{\link{NullOrCharacter-class}},
  \code{\link{NullOrInteger-class}}, \code{\link{conrpc}},
  \code{\link{rpcpost}}, \code{\link{startbtc}},
  \code{\link{stopbtc}}
}
\concept{bitcoind functions}
