% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CONRPC-class}
\alias{CONRPC-class}
\title{The CONRPC class}
\description{
S4-class for curl connections to RPC-JSON.
}
\details{
The slots \code{rpcuse} and \code{rpcpwd} are required in the call
to \code{curl}. Furthermore, the fully qualified path to
\code{bitcoin.conf} (slot \code{config}) is required for starting
and stopping \code{bitcoind} as daemon.
}
\seealso{
Other bitcoind functions: \code{\link{ANSRPC-class}},
  \code{\link{NullOrCharacter-class}},
  \code{\link{NullOrInteger-class}}, \code{\link{conrpc}},
  \code{\link{rpcpost}}, \code{\link{startbtc}},
  \code{\link{stopbtc}}
}
\concept{bitcoind functions}
