% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoind.R
\name{conrpc}
\alias{conrpc}
\title{Extracting Configuration Settings}
\usage{
conrpc(conf.file)
}
\arguments{
\item{conf.file}{\code{character}, the fully qualified path.}
}
\value{
An S4-object of class \code{CONRPC}.
}
\description{
This function extracts information from the configuration
file \code{bitcoin.conf} with respect to the options  \code{rpcuser}
and \code{rpcpassword}.
}
\seealso{
CONRPC-class, startbtc, stopbtc
}
\author{
Bernhard Pfaff
}
