% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilFuncs.R
\name{date2int}
\alias{date2int}
\title{Convert date/time to integer}
\usage{
date2int(x)
}
\arguments{
\item{x}{\code{POSIXct}, date/time object.}
}
\value{
\code{integer}
}
\description{
This function returns the associated \code{integer} time
for a given date/time object (coercible as \code{POSIXct} object.
}
\examples{
d <- "2017-03-15"
date2int(d)
}
\seealso{
Other UtilityFuncs: \code{\link{bkfee}},
  \code{\link{blockattime}}, \code{\link{blockstats}},
  \code{\link{int2date}}, \code{\link{intMaxDay}},
  \code{\link{intMinDay}}, \code{\link{intRangeDay}},
  \code{\link{intRangePeriod}}, \code{\link{timeofblock}},
  \code{\link{txfee}}, \code{\link{txids}},
  \code{\link{txinids}}, \code{\link{txstats}},
  \code{\link{utxoage}}, \code{\link{utxotype}},
  \code{\link{utxovalue}}
}
\author{
Bernhard Pfaff
}
\concept{UtilityFuncs}
