% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getblockchaininfo}
\alias{getblockchaininfo}
\title{RPC-JSON API: getblockchaininfo}
\usage{
getblockchaininfo(obj)
}
\arguments{
\item{obj}{object of class \code{CONRPC}.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returns an object containing various state info regarding blockchain processing.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getblockchaininfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{getbestblockhash}},
  \code{\link{getblock}}
}
\author{
Bernhard Pfaff
}
