% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRpcs.R
\name{gethelp}
\alias{gethelp}
\title{RPC-JSON API: Help}
\usage{
gethelp(conobj, rpcname = "")
}
\arguments{
\item{conobj}{object of class \code{CONRPC}.}

\item{rpcname}{\code{character}, name of RPC function.}
}
\value{
A coerced \code{list} object from RPC-JSON API.
}
\description{
Returning information about RPC functions.
}
\references{
\url{https://bitcoin.org/en/developer-reference#gethelp},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Control RPCs: \code{\link{getblockchaininfo}},
  \code{\link{getinfo}}, \code{\link{getnetworkinfo}},
  \code{\link{getwalletinfo}}
}
\author{
Bernhard Pfaff
}
