% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRpcs.R
\name{getinfo}
\alias{getinfo}
\title{RPC-JSON API: getinfo}
\usage{
getinfo(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returning information about bitcoin configuration and settings.
}
\details{
WARNING: getinfo is deprecated and will be fully removed in 0.16.
Projects should transition to using getblockchaininfo, getnetworkinfo, and
getwalletinfo before upgrading to 0.16.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Control RPCs: \code{\link{gethelp}},
  \code{\link{getwalletinfo}}
}
\author{
Bernhard Pfaff
}
\concept{Control RPCs}
