% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockchainRpcs.R
\name{getmempoolentry}
\alias{getmempoolentry}
\title{RPC-JSON API: getmempoolentry}
\usage{
getmempoolentry(con, txid)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}

\item{txid}{\code{character}, the transaction id (must be in mempool).}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns mempool data for given transaction.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getmempoolentry},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Blockchain RPCs: \code{\link{decodescript}},
  \code{\link{getbestblockhash}},
  \code{\link{getblockchaininfo}},
  \code{\link{getblockcount}}, \code{\link{getblockhash}},
  \code{\link{getblockheader}}, \code{\link{getblock}},
  \code{\link{getchaintips}},
  \code{\link{getchaintxstats}},
  \code{\link{getdifficulty}},
  \code{\link{getmempoolancestors}},
  \code{\link{getmempooldescendants}},
  \code{\link{getmempoolinfo}},
  \code{\link{getrawmempool}}, \code{\link{gettxoutproof}},
  \code{\link{gettxoutsetinfo}}, \code{\link{gettxout}},
  \code{\link{preciousblock}},
  \code{\link{pruneblockchain}}, \code{\link{verifychain}},
  \code{\link{verifytxoutproof}}
}
\author{
Bernhard Pfaff
}
\concept{Blockchain RPCs}
