% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkRpcs.R
\name{getnetworkinfo}
\alias{getnetworkinfo}
\title{RPC-JSON API: getnetworkinfo}
\usage{
getnetworkinfo(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returns an object containing various state info regarding P2P networking.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getnetworkinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Network RPCs: \code{\link{addnode}},
  \code{\link{clearbanned}}, \code{\link{disconnectnode}},
  \code{\link{getaddednodeinfo}},
  \code{\link{getconnectioncount}},
  \code{\link{getnettotals}}, \code{\link{getpeerinfo}},
  \code{\link{listbanned}}, \code{\link{ping}},
  \code{\link{setnetworkactive}}
}
\author{
Bernhard Pfaff
}
\concept{Network RPCs}
