% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlRpcs.R
\name{getwalletinfo}
\alias{getwalletinfo}
\title{RPC-JSON API: getwalletinfo}
\usage{
getwalletinfo(con)
}
\arguments{
\item{con}{object of class \code{CONRPC}.}
}
\value{
A S4-object of class \code{ANSRPC}.
}
\description{
Returning information about bitcoin wallet.
}
\references{
\url{https://bitcoin.org/en/developer-reference#getwalletinfo},
\url{https://bitcoin.org/en/developer-reference#remote-procedure-calls-rpcs}
}
\seealso{
Other Control RPCs: \code{\link{gethelp}},
  \code{\link{getinfo}}
}
\author{
Bernhard Pfaff
}
\concept{Control RPCs}
