% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoind.R
\name{startbtc}
\alias{startbtc}
\title{Start bitcoind server process}
\usage{
startbtc(confbtc)
}
\arguments{
\item{confbtc}{\code{CONRPC} object, returned from \code{conrpc()}.}
}
\value{
\code{NULL}
}
\description{
This function does start the bitcoind-server process.
It should only be called when no suitable RPC-JSON
process is running
}
\details{
The process is started by calling \code{system()}.
Hereby, the options: \code{rpcuser}, \code{rpcpassword} and
\code{conf} are used in the call to \code{bitcoind}.
}
\seealso{
Other bitcoind functions: \code{\link{ANSRPC-class}},
  \code{\link{CONRPC-class}},
  \code{\link{NullOrCharacter-class}},
  \code{\link{NullOrInteger-class}}, \code{\link{conrpc}},
  \code{\link{rpcpost}}, \code{\link{stopbtc}}
}
\author{
Bernhard Pfaff
}
\concept{bitcoind functions}
