% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoind.R
\name{stopbtc}
\alias{stopbtc}
\title{Stop bitcoind server process}
\usage{
stopbtc(confbtc)
}
\arguments{
\item{confbtc}{\code{CONRPC} object, returned from \code{conrpc()}.}
}
\description{
This function stops a running bitcoind process.
It calls \code{bitcoin-cli stop} \emph{via} the
R function \code{system()}.
}
\seealso{
Other bitcoind functions: \code{\link{ANSRPC-class}},
  \code{\link{CONRPC-class}},
  \code{\link{NullOrCharacter-class}},
  \code{\link{NullOrInteger-class}}, \code{\link{conrpc}},
  \code{\link{rpcpost}}, \code{\link{startbtc}}
}
\author{
Bernhard Pfaff
}
\concept{bitcoind functions}
